requirements.txt
pandas
sklearn
matplotlib
keras==2.3.1
pyyaml
tqdm
tensorflow==1.13.1
h5py==2.10.0
pydot # to plot model using keras
#sudo apt install python-pydot python-pydot-ng graphviz
#scipy==1.3.0
#numpy==1.17.3